add force option > chocolatey-install.ps1

+++ chocolatey-install.ps1
@@ -1,5 +1,9 @@
 #!pwsh
 
+param(
+    [switch]$Force
+)
+
 $ExecutionPolicy = Get-ExecutionPolicy
 if ($ExecutionPolicy -eq 'RemoteSigned') {
     Set-ExecutionPolicy RemoteSigned
@@ -10,4 +14,9 @@ if (!(Get-Command choco -ea SilentlyContinue)) {
     Invoke-Expression ((New-Object System.Net.WebClient).DownloadString('https://chocolatey.org/install.ps1'))
 }
 
-choco install chocolatey-package.config --yes
+if ($Force) {
+    choco install chocolatey-package.config --force --yes
+}
+else {
+    choco install chocolatey-package.config --yes
+}
